Sync Prettier options to buffer-local variables
This syncs various Prettier options, such as `printWidth` and
`useTabs`, to the corresponding Emacs variables, such as
`js-indent-level`.

Emacs variables are set for a number of major modes for languages that
Prettier supports either directly or through plug-ins.

Buffer-local variables are removed or restored to their previous value
when `prettier-js-mode` is disabled, unless the variable was changed
since `prettier-js-mode` was enabled.